requirements.txt
streamlit==1.39.0
langchain==0.3.0
langchain-community==0.3.0
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langchain-huggingface==0.1.0
faiss-cpu==1.8.0.post1
pypdf==5.0.1
python-dotenv==1.0.1
tavily-python==0.5.0